Pushed zip and war files of the API
This CRUD API is created to demonstrate the different operations on the records of the music tracks. The API functions as a simple service that can retrieve information from the database about the details of all tracks, insert new track record, update information and delete a track record. Thus, the functionalities of the API include: 1] Retrieve the title, artist, album and genre details of all the tracks, 2] Insert a new track record, 3] Update track information, and 4] Delete a track record.
This API has been deployed on the NEXEN Cloud platform.